libgen: array subscript is of type 'char'
../common/reg_step.c:298:12: error: array subscript is of type 'char' [-Werror,-Wchar-subscripts]
                        braslist[*ep++] = lp;
                                ^~~~~~
../common/reg_step.c:302:12: error: array subscript is of type 'char' [-Werror,-Wchar-subscripts]
                        braelist[*ep++] = lp;
                                ^~~~~~
../common/reg_step.c:418:19: error: array subscript is of type 'char' [-Werror,-Wchar-subscripts]
                        bbeg = braslist[*ep];
                                       ^~~~
../common/reg_step.c:419:23: error: array subscript is of type 'char' [-Werror,-Wchar-subscripts]
                        ct = (int)(braelist[*ep++] - bbeg);
                                           ^~~~~~
../common/reg_step.c:428:19: error: array subscript is of type 'char' [-Werror,-Wchar-subscripts]
                        bbeg = braslist[*ep];
                                       ^~~~
../common/reg_step.c:429:23: error: array subscript is of type 'char' [-Werror,-Wchar-subscripts]
                        ct = (int)(braelist[*ep++] - bbeg);
                                           ^~~~~~
6 errors generated.

+++ usr/src/lib/libgen/common/reg_step.c
@@ -295,11 +295,11 @@ _advance(char *lp, char *ep)
 			continue;
 
 		case CBRA:
-			braslist[*ep++] = lp;
+			braslist[(int)*ep++] = lp;
 			continue;
 
 		case CKET:
-			braelist[*ep++] = lp;
+			braelist[(int)*ep++] = lp;
 			continue;
 
 		case MCCHR | RNGE:
@@ -415,8 +415,8 @@ _advance(char *lp, char *ep)
 			goto star;
 
 		case CBACK:
-			bbeg = braslist[*ep];
-			ct = (int)(braelist[*ep++] - bbeg);
+			bbeg = braslist[(int)*ep];
+			ct = (int)(braelist[(int)*ep++] - bbeg);
 
 			if (ecmp(bbeg, lp, ct)) {
 				lp += ct;
@@ -425,8 +425,8 @@ _advance(char *lp, char *ep)
 			return (0);
 
 		case CBACK | STAR:
-			bbeg = braslist[*ep];
-			ct = (int)(braelist[*ep++] - bbeg);
+			bbeg = braslist[(int)*ep];
+			ct = (int)(braelist[(int)*ep++] - bbeg);
 			curlp = lp;
 			while (ecmp(bbeg, lp, ct))
 				lp += ct;
